refactor: remove mpf from sdk
