+++ demo/midgard-sdk/src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from "./common.js";
 export * from "./ledger-state.js";
 export * from "./linked-list.js";
-export * from "./mpt.js";
 export * from "./state-queue.js";
